Removed deprecated features

+++ src/app/src/main/java/com/jm/blogitz/BlogPagerActivity.java
@@ -2,9 +2,6 @@ package com.jm.blogitz;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.content.pm.ResolveInfo;
-import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.net.Uri;
 import android.os.Bundle;
@@ -16,10 +13,11 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.content.FileProvider;
+import androidx.core.content.res.ResourcesCompat;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentStatePagerAdapter;
-import androidx.viewpager.widget.ViewPager;
+import androidx.fragment.app.FragmentActivity;
+import androidx.viewpager2.adapter.FragmentStateAdapter;
+import androidx.viewpager2.widget.ViewPager2;
 
 import com.jm.blogitz.models.Blog;
 
@@ -29,8 +27,30 @@ import java.util.UUID;
 
 public class BlogPagerActivity extends AppCompatActivity {
     private static final String EXTRA_BLOG_ID = "com.jm.blogitz.blog_id";
-    private ViewPager viewPager;
     private List<Blog> blogs;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_blog_pager);
+
+        assert getSupportActionBar() != null;
+        ActionBar actionBar = getSupportActionBar();
+        actionBar.setDisplayHomeAsUpEnabled(true);
+        int color = ResourcesCompat.getColor(getResources(), R.color.background_dark, getTheme());
+        ColorDrawable colorDrawable = new ColorDrawable(color);
+        actionBar.setBackgroundDrawable(colorDrawable);
+
+        UUID blogId = (UUID) getIntent().getSerializableExtra(EXTRA_BLOG_ID);
+        this.blogs = BlogLab.get(this).getBlogs();
+
+        ViewPager2 viewPager = findViewById(R.id.activity_blog_pager_view_pager);
+        FragmentStateAdapter pagerAdapter = new ScreenSlidePagerAdapter(this);
+        viewPager.setAdapter(pagerAdapter);
+
+        int blogIndex = this.findBlogIndex(blogId);
+        viewPager.setCurrentItem(blogIndex);
+    }
 
     public static Intent newIntent(Context packageContext, UUID blogId) {
         Intent intent = new Intent(packageContext, BlogPagerActivity.class);
@@ -50,66 +70,51 @@ public class BlogPagerActivity extends AppCompatActivity {
         BlogLab blogLab = BlogLab.get(this);
 
         Blog blog = blogLab.getBlog(blogId);
+        blog.validateBlog(this);
 
         File photoFile = blogLab.getPhotoFile(blog);
         Uri photoFileUri = FileProvider.getUriForFile(this, "com.jm.blogitz.fileprovider", photoFile);
         String title = blog.getTitle();
         String body = blog.getBody();
 
-        switch (item.getItemId()) {
-            case R.id.publish_menu_item:
-                dispatchTwitterPublishIntent(title, body, photoFileUri);
-                return true;
-            case R.id.share_via_email_menu_item:
-                dispatchSendEmailIntent(title, body, photoFileUri);
-                return true;
-            case R.id.delete_menu_item:
-                blogLab.deleteBlog(blog);
-                dispatchMoveBackIntent();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        int itemId = item.getItemId();
+        if (itemId == R.id.publish_menu_item) {
+            dispatchTwitterPublishIntent(title, body, photoFileUri);
+            return true;
+        } else if (itemId == R.id.share_via_email_menu_item) {
+            dispatchSendEmailIntent(title, body, photoFileUri);
+            return true;
+        } else if (itemId == R.id.delete_menu_item) {
+            blogLab.deleteBlog(blog);
+            dispatchMoveBackIntent();
+            return true;
+        } else {
+            return super.onOptionsItemSelected(item);
         }
     }
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_blog_pager);
-
-        assert getSupportActionBar() != null;
-        ActionBar actionBar = getSupportActionBar();
-        actionBar.setDisplayHomeAsUpEnabled(true);
-        ColorDrawable colorDrawable = new ColorDrawable(Color.parseColor("#141218"));
-        actionBar.setBackgroundDrawable(colorDrawable);
-
-        UUID blogId = (UUID) getIntent().getSerializableExtra(EXTRA_BLOG_ID);
-        this.viewPager = findViewById(R.id.activity_blog_pager_view_pager);
-        this.blogs = BlogLab.get(this).getBlogs();
-
-        FragmentManager manager = getSupportFragmentManager();
-        this.viewPager.setAdapter(new FragmentStatePagerAdapter(manager) {
-            @NonNull
-            @Override
-            public Fragment getItem(int position) {
-                Blog blog = blogs.get(position);
-                return BlogFragment.newInstance(blog.getId());
-            }
-
-            @Override
-            public int getCount() {
-                return blogs.size();
-            }
-        });
-
-        int blogIndex = this.findBlogIndex(blogId);
-        this.viewPager.setCurrentItem(blogIndex);
+    public boolean onSupportNavigateUp() {
+        this.dispatchMoveBackIntent();
+        return true;
     }
 
-    @Override
-    public boolean onSupportNavigateUp() {
-        finish();
-        return true;
+    private class ScreenSlidePagerAdapter extends FragmentStateAdapter {
+        public ScreenSlidePagerAdapter(FragmentActivity fa) {
+            super(fa);
+        }
+
+        @NonNull
+        @Override
+        public Fragment createFragment(int position) {
+            Blog blog = blogs.get(position);
+            return BlogFragment.newInstance(blog.getId());
+        }
+
+        @Override
+        public int getItemCount() {
+            return blogs.size();
+        }
     }
 
     private void dispatchTwitterPublishIntent(String title, String body, Uri fileUri) {
